Fix a byte-order bug, implement new instructions, fix wrapping add/sub

+++ src/gameboy/constants.rs
@@ -1,10 +1,10 @@
 pub const WRAM_SIZE: usize = 8192;
 
 // TODO: Switchable ROM
-pub const ROM_START: usize = 0x0000;
-pub const ROM_END: usize = 0x3FFF;
+pub const ROM_START: u16 = 0x0000;
+pub const ROM_END: u16 = 0x3FFF;
 
 // TODO: VRAM etc
 
-pub const WRAM_START: usize = 0xC000;
-pub const WRAM_END: usize = 0xDFFF;
+pub const WRAM_START: u16 = 0xC000;
+pub const WRAM_END: u16 = 0xDFFF;
